[FIX] Cheering pet doesn't set it as managed pet for the day
